avoid floating point math
reservation right now is 25%, which means that the requested size of a
region can be divided by 4. avoid floating point math where possible.

change the reservation percentage stored with the region to an enum,
where the only value is 25%. this limits what can be done with manual
database edits, and restricts what the Region::reserved_size function
has to guard against.

it'd be nice if Region::new was a test-only function but the crate
doesn't have the same idea of a integration test feature.