ctest2 some structs in illumos-utils
this "works" but seems more brittle than it's worth at the moment.
`ctest2` only supports code that compiles under edition 2015, so among
other things we can't just test against all public items in
`illumos-utils` in one go.

This really points towards having a `-sys` crate binding types and
headers, which illumos-utils uses in interacting with the system. We've
talked about doing this anyway, so this is more of a demo to carry
forward when we do that.

The structs we'd be testing now either have been stable for 20 years or
are precisely described in man pages (implying forward compatibility),
so the marginal value of landing this now with testing is.. not fantastic.