[1/n] move db schema into its own crate (#7880)
While working on Diesel-related code, I noticed that schema generation is both
_very_ large (~450k lines of code currently), and also can be separated out
into its own crate. Spent a couple of hours doing so and I think I quite like
the result! The most important thing here is that the schema code is completely
independent of the rest of Omicron, so it can be compiled in parallel with
crates like omicron-common.

For me, a full `cargo check --all-targets` in Omicron goes from 2m04s to 1m51s,
which is a really nice speedup. Incremental builds should also be faster due to
the parallelization and reduction of crate size. (But this doesn't really touch
nexus-db-queries, which is the worst offender at the moment.)