[2/n] [nexus-db-schema] make HAS_STATIC_QUERY_ID true for all enums (#7897)
This is meant to be false if and only if types don't match 1:1 to
queries, e.g. with CTEs. These types do match 1:1 to queries, so using
`HAS_STATIC_QUERY_ID` should be fine.

There's a small possibility some code was inadvertently depending on
`HAS_STATIC_QUERY_ID` being false, but tests should hopefully catch
that.

Depends on #7880.