Insert project filters into OxQL query AST instead of raw string (#7904)
Closes #7532. Instead of manipulating the query as a string to add the
filters at the end, pass the silo and project IDs to the oximeter
client, which has the parsed query AST on hand, and let it do the
manipulation.

In case of simple queries, we just insert the filters after the initial
`get` operation. For "grouped" queries (like `{ get a:b; get c:d |
filter timestamp > @now() } | filter x > y`), the query is a tree, so we
recurse down it and insert the filters in every leaf node, which is a
simple query.