sled-agent: clean up and reorder VMM termination logic
Try to clean up sled-agent's logic for terminating Propolis zones.
Specifically, change `InstanceRunner::terminate` and its callees so that
they no longer call `InstanceTicket::deregister` before the runner gets
a chance to publish a terminal VMM state to Nexus. This fixes a race in
which Nexus could call sled-agent, get a "VMM gone" response, and move
the relevant VMM to Failed when sled-agent actually intended to move it
to Destroyed. 

Do some housekeeping to (try to) make this code easier to reason about:

- Rename various types and routines to make their purposes and semantics
  a little clearer. (In particular, try not to have so many routines
  named "terminate" or variations thereon.)
- Clean up some code and comments that are obsolete now that the
  instance monitor and instance runner tasks pass messages instead of
  sharing state.
- Make explicit termination requests specify who is responsible for
  updating the terminated VMM's state: sometimes it's sled-agent (e.g.
  when terminating a Propolis because its U.2 is gone), but sometimes
  it's Nexus (e.g. when unwinding an instance start saga). This allows
  the instance runner to avoid updating a VMM that Nexus expects to be
  able to control.
- Remove `InstanceTicket::deregister` and move its implementation into
  `InstanceTicket::drop`. This prevents the runner from sending a
  "deregister me" message "too early"; it can still forget to send a
  VMM's terminal state to Nexus, but the structure of the main runner
  loop should help ensure that this last update always gets sent.