[nix] Fix Nix flake to track changes from #7907
PR #7907 broke the Nix flake, as it removed the
`tools/dendrite_openapi_version` file and replaced it with
`tools/dendrite_version`. The flake was previously using
`tools/dendrite_openapi_version` to determine both the Dendrite OpenAPI
commit and the Dendrite stub commit to download. This commit changes it
to use the new `dendrite_version` file to determine the commit for
downloading the Dendrite stub, and removed the code for downloading the
Dendrite OpenAPI document.