manually manage instance ticket lifetimes
Instead of making a VM's InstanceTicket a member of InstanceRunner and
relying on it being dropped when `InstanceRunner::run` returns, pass the
ticket as a parameter to `run` and drop it explicitly at the appropriate
moment. This is needed to avoid deadlocking the runner when it wants to
exit (it needs to see that the sender side of its request queue is
dropped, and this requires the controlling instance manager to release
the relevant instance).