oxlog: Allow filtering by file mtime (#7905)
On dogfood we now have several years of log archives, and a naive grep
over all of the archived logs can take minutes to complete. A preemptive
filter in `oxlog` to find only the logs within the rough time period
we're interested in will make this kind of search much more efficient.

Add `--after` and `--before` flags to the CLI to filter logs based on
their `mtime`. This uses `jiff` so that
relative timestamps like '6 hours ago' or '-1 hour' can be passed, in
addition to precise datetimes like '2025-04-02T12:21:09Z'.

Implement these new CLI arguments outside of the `FilterArgs` struct to
ensure they don't bypass the requirement for users to explicitly specify
one of the mandatory filter flags (`--current`, `--archived`, or
`--extra`).