Improvements to time synchronisation in the face of connectivity problems. (#7675)
For consistency of time within the rack we must guard against there ever
being two authoritative sources of time. We currently have two
(admittedly
edge) cases where this can occur.

1) When, some time after everything is synchronised, one or both of the
   boundary NTP servers loses upstream connectivity, but continues to
   advertise at the same stratum as its clock begins to drift.
2) If both boundary NTP servers lose connectivity and fall back to their
   local clocks, advertising them with stratum 10, they will both be
   authoritative but with potentially different times.

This change addresses both of these by updating the NTP server
configuration
in a number of ways.

1) Adding each boundary server as a source to the other;
2) Configuring the boundary "local" sources with the "orphan" flag that
causes selection of just one as authoritative if both are in that mode;
3) Configuring RSS sources with a new "failfast" flag that causes them
to
be discounted quickly (marked as "unselectable") once they are
considered
unreachable, instead of waiting for their "distance" to degrade over
time;
4) Adjusting the root dispersion decay rate from chrony's default of
1µs/s
(versus RFC recommended default of 15µs/s) up to 60µs/s to achieve
faster
   source reselection.

This is partially in response to the problems encountered in
https://github.com/oxidecomputer/omicron/issues/7534