[sled-agent] Ensure Support Bundles datasets are mounted (#7938)
Related to https://github.com/oxidecomputer/omicron/pull/7887 - we have
seen that datasets nested within
encrypted datasets are not automatically mounted on reboot.

To mitigate this issue: whenever we perform an operation which attempts
to read the dataset, ensure
that it is mounted. To do this operation, some minor refactors were made
to mounting-related code.

Fixes https://github.com/oxidecomputer/omicron/issues/7884