Reduce regions_hard_delete contention (#7960)
Recently, regions_hard_delete changed to use a CTE to update the
size_used column of the crucible_dataset table. Unfortunately this had
the side effect of causing a massive amount of contention: the CTE would

1) delete rows from the regions table
2) read from the regions table during the CTE
3) update the size_used column for _all_ crucible_dataset rows

This would almost certainly cause each invocation of the CTE to contend
with each other, as seen when doing disk deletes in parallel.

This commit changes the CTE to:

1) delete rows from the regions table, returning the affected datasets
2) read from the regions table during the CTE
3) update the size_used column for affected crucible_dataset rows only.

This was tested by using terraform to create and tear down 90 disks,
with a parallelism setting of 10, 20, and 30. Before this change, this
would not work as Nexus would inevitably return 500s.

Fixes #7952