Reconcile differences in switch port api types
Switch port settings create params takes a hashmap of
link name -> configuration value. However, we return a struct
that has the configuration value and link name inside of a
single type. We are updating this to be the same for create params
and returned types.

This change accomplishes two things:
* The API is more consistent for users
* The generated go client doesn't seem to work properly with the
  hash map style params. This change allows us to properly create
  settings using the go client.