Resolve Nexus and renew oximeter lease in the same operation (#7990)
- Oximeter producers are expected to periodically renew their leases
with Nexus, to help prune producers that have disappeared. Previously,
the method responsible for this first looked up Nexus in DNS, and _then_
attempted to renew a lease, backing off on failures to reach Nexus. This
meant it's possible to look up a Nexus address; expunge that Nexus zone;
and then continue to use the now-defunct address forever. This changes
the method to both resolve Nexus and attempt to register in the same API
call, with backoff between each of them as a unit.
- Fixes #6158