omdb: add facility for abandoning a saga (#7791)
Add an `Abandoned` saga state. This state disqualifies a saga from being
picked up by Nexus saga recovery. (A running saga will continue running
if it is Abandoned, and continued saga execution may end up clobbering
the Abandoned state entirely.) Add an `omdb` subcommand to move a saga
to this state (and refactor a bit to avoid duplicating code with the
`inject-error` subcommand).