Trust quorum: reconfiguration and commit behavior
This PR adds further functionality to the sans-io trust quorum protocol.
Configurations can now be committed via `Node::commit_reconfiguration`.
For each reconfiguration attempt made on top of a committed
configuration, the rack secret for the last committed reconfiguration
will be reconstructed after retreiving a threshold of shares from
members of that configuration. At this point this "old" rack secret will
be encrypted with a key derived from the rack secret for the current
configuration being coordinated and included as necessary in prepare
messages sent out during coordination.

The property based test for coordinator behavior has been expanded to
include support for this functionality, as well as to allow dropping
messages between nodes if such an action is generated. The bulk of
this PR lies in the test code, and it has been restructured to handle
multiple reconfigurations and commits. This has led to the tracking of
shares across non-existent test nodes, and enhancements to the model.

Additionally, a small change was made to copy some of the errors out of
`validators.rs` and into their own file.