oxlog: Glob match zone names
`oxlog logs` currently requires a precise match for the zone name. This
is inconvenient when attempting to search across multiple zones with
similar names. For example, when searching all Crucible downstairs, or
using Pilot to query Nexus logs on multiple sleds.

Use the `glob` crate to take a pattern to match against the zone name
for the `services` and `logs` subcommands, potentially allowing multiple
zones to be returned. If more than one zone is matched, then the logs
and services are sorted separately.

Zones with a large number of archived logs may take an extended period
to walk and sort, and when matching against multiple zones this becomes
uncomfortably slow. To claw back some performance, use Rayon to find
files and sort in parallel. We limit the size of its thread pool to a
maximum of eight threads to avoid soaking up all available threads on a
sled. On smaller machines we fall back to the number of logical threads.

Example usage:

  $ oxlog services 'oxz_crucible_[!p]*' # All non-pantry Crucible zones
  $ oxlog logs 'oxz_nexus_*' --current  # The nexus zone