Trust Quorum: Start implementing `Node` (#7929)
This code builds upon #7859, which itself builds upon #7891. Those need
to be merged in order first.

A `Node` is the sans-io entity driving the trust quorum protocol. This
PR starts the work of creating the `Node` type and coordinating an
initial configuration.

There's a property based test that generates an initial `ReconfigureMsg`
that is then used to call `Node::coordinate_reconfiguration`, which will
internally setup a `CoordinatorState` and create a bunch of messages to
be sent. We verify that a new `PersistentState` is returned and that
messages are sent.

Some of validation code is also tested. We test a subset of error
conditions, but ensure all the error check methods themselves actually
get called among the checks we choose.

---------

Co-authored-by: Rain <rain@oxide.computer>