Move CRDB initialization from sled-agent to cockroach-admin (#8026)
The main change in this PR is removing the sled-agent `POST
/cockroachdb` endpoint and adding the cockroach-admin `POST
/cluster/init` endpoint. This endpoint is only used during rack setup,
and is what runs both `cockroach init` and `dbinit.sql`.

There's one semantic change: the sled-agent handler also ran
`dbwipe.sql`, and I dropped that from the cockroach-admin handler. I
don't think there's any need to wipe: `dbinit.sql` runs inside a
transaction, so there shouldn't be any case where it partially succeeds
such that a retry would need to wipe before rerunning `dbinit.sql`. And
it seems unreasonably dangerous to have an endpoint that can be `POST`'d
to that wipes the entire db (which is currently the case in
sled-agent!).

A couple other side benefits:

* sled-agent had to zlogin to the cockroach zone to do these operations;
cockroach-admin doesn't since it's running in the zone already.
* This removes a thorny "where should this sled-agent code live" case on
the upcoming reconciler branch.