Support bundles should include log files (#7973)
Take two of adding log files into support bundles.

This PR leverages oxlog to find all of the zones/services on a sled and
then through a series of zfs snapshots will collect those logs and
expose them as a zip file that nexus will slurp down and stick into the
support bundle that is being collected.