[4/n] [installinator] split fetching and progress reporting traits (#8038)
While working on better progress reporting, I realized that fetching and progress reporting really are quite different, and are best represented through different traits.

This removes several panics and implicit unused behaviors, so it acts as a really nice cleanup on top of facilitating the next few commits in the series.

There are no functional changes in this PR -- it's just refactoring.