sled-diagnostics find_log_in_snapshot needs to be more robust (#8048)
While testing support bundle collection on london I noticed that we were
not collecting archived logs properly.
Upon investigation I found the following in the sled-agent logs:
```
20:30:23.070Z ERRO SledAgent: found log file that is not a file, skipping over it but this is likely a programming error
    file = sled-diagnostics/src/logs.rs:494
    logfile = /pool/ext/110131b4-7bde-4866-b37e-bd9e3ebcbdf3/crypt/debug/.zfs/snapshot/sled-diagnostics-mUAUCJAKFhLe/oxide-zone-network-setup:default.log.536112088
    sled_id = 7c3a1cd5-3672-4d91-b4d4-064c269a7d96
```

This PR makes finding the log we are operating on in a snapshot more
robust by reconstructing the final log path via the ZFS snapshot
mountpoint taking into account the initial log file path. As a bonus
this cleans things up and removes some guess work that was previously
done with delegated datasets and zone names.