[nexus] track "intended" instance states
When a request to a sled-agent to change the state of an instance fails
in such a way that indicates that the instance is no
longer present on that sled, the instance record is moved to the
`Failed` state. If an instance is in the `Failed` state and auto-restart
is enabled for that instance, then the control plane will always attempt
to restart it.

Both of these behaviors are fine in a vacuum, but in conjunction, they
result in somewhat surprising behavior in the case where an instance
which has failed is asked to stop. If the request that the sled-agent
stop the instance fails in a way that indicates that the instance is
gone, then Nexus will move the instance to `Failed` and attempt to
restart it.[^1] This is, to use a technical term, wacky. The user asked
that an instance not be running, we went and tried to stop it and
discovered that it was, in fact, not running, and then we...decided to
start it again? That's literally the opposite of the intended behavior,
from the user's perspective. I described this in #6809.

This commit fixes this (and lays the groundwork for some other future
improvements to instance state management). The approach is as follows:

1. Introduce a new "intended state" column to database instance records,
   tracking the most recently requested state of the instance. This
   column will only record "resting" states that can be requested through
   the API: `Stopped`, `Running`, and `Destroyed`.
2. When handling API requests that change the state of an instance,
   immediately update the `intended_state` of the instance record before
   doing anything else (starting sagas, sending requests to sled-agents,
   etc.)
3. When deciding whether to automatically restart a `Failed` instance,
   consider the intended state as well as other variables, and only
   restart it if it is supposed to be `Running`

Presently, this tracking of requested states is only used for
determining whether to auto-restart an instance. However, it may be
useful for at laest one other thing that kinda bothers me: when an
instance is created with the `start: true` field in the `InstanceCreate`
params, the API endpoint runs an `instance_create` saga and then, once
it completes, also runs an `instance_start` saga. Both operations are
performed in sagas, so if the Nexus instance that is handling the API
request dies unexpectedly, the user's intent should be carried
out...except this isn't _quite_ the case. If the Nexus dies while
running the `instance_create` saga, or after the `instance_create` saga
has completed but before the `instance_start` saga has been created and
written to the database, then the instance will be created, but the user
intent that the instance be *started* will not be carried out.

This is a bit unfortunate, especially when the instance is created by
automation against the Oxide API. A human starting an instance in the
console may just notice that the instance has been created and is now in
the `Stopped` state, and click on the "start" button, and just be a
little bit annoyed that their instance didn't start automatically when
they clicked a checkbox saying it should. But, automation against the
API might somewhat reasonably expect that if an instance-create request
has `start: true`, the instance will be started once it is created, and
not try to start it again, resulting in the instance never actually
starting. We could fix this using the tracked instance intended states
as well. The completion of the `instance-create` saga could "chain" into
an `instance-start` saga if the instance's intended state is `Running`,
or an RPW could ensure that an `instance-start` saga exists for all
instances with that intended state. I haven't implemented that in this
PR, but this change lays the groundwork for that as well.

[^1]: Provided that auto-restart is configured for that instance.