update to track sp-sim changes from #8055
This necessitates moving the construction of `SpUpdate` earlier in the
sim Gimlet initialization, so that the ereport state can ask it for
Hubris version metadata. Now, it has to be constructed in
`Gimlet::spawn` and passed in to `UdpTask::new` (and thus
`Handler::new`), rather than constructed isnide `Handler::new`, so that
the ereport state can also be passed in to `Handler::new`. @dap, let me
know what you think of this --- if you base additional changes on code
with the current factoring, it will probably make the eventual merge
conflicts between our branches much more unpleasant, so maybe we can
pull this out and land it separately, and have my branch depend on that?
Let me know what you think!