Remove alarms from persistent state
Instead we return them wherever they can arise. This has a couple
of benefits.

* Higher level code can stop accepting requests when it sees an alarm to
prevent endless logging and raise an alert to Nexus.
* After support resolves the issue there is not necessarily any reason
to manually mutate persistent state, unless that was the cause of the
alarm.

This also allows more straightforward/rusty error handling. Persistent
state is now only returned in success cases.