Introduce sled-agent-config-reconciler skeleton (#8063)
This PR adds a new `sled-agent-config-reconciler` crate. The majority of
its public API is present here, but most of the guts of it is stubbed
out via `unimplemented!()`. This is the first PR of many in a big chunk
of work, and I'm not sure how best to divide it, but my current plan is:

1. This PR
2. A draft PR that shows how this public API is integrated into
sled-agent
3. A series of (relatively) smaller PRs that flesh out all the
`unimplemented!()` bits inside the reconciler
4. (Optional) PRs to finish any lingering integration work not covered
by 2

This PR is safe to land on `main`: the only changes to sled-agent proper
are that the `dump_setup` module has moved to this crate (with some very
minor changes that don't affect behavior), so sled-agent calls it from
this crate instead of from its own submodule. The rest of the crate with
all the `unimplemented!()` paths is not called by anyone.

The PRs in group 3 will similarly be safe to land: they will only flesh
out the `unimplemented!()`-but-not-called code paths.

PR 2 (which I'll open shortly) will not pass tests until all the PRs
from group 3 have landed, but I'd like to open them in this order to get
eyes on that integration work, because it may inform changes we want to
make to the config-reconciler crate's API (which would be easier now
than after landing everything in group 3).