[sled-agent-config-reconciler] Flesh out LedgerTask (#8075)
This implements the task responsible for accepting new
`OmicronSledConfig`s, validating them (generation isn't going backwards
and they don't reference artifacts we don't have), and writing them to
the config datasets of the M.2s. It also includes code for converting
the legacy triple of disk/dataset/zone configs into the new combined
config (which is a lossless operation).

In terms of code size, this PR is smaller than it looks: about half the
additions are test data (JSON), and of the remaining additions close to
half of them are tests. This is safe to merge into `main`, because it's
replacing `unimplemented!()` bits that are not currently used with
implemented bits that are still not currently used.

It introduces a couple potential panics in sled-agent startup for very
weird edge cases:

1. We don't have a ledgered `OmicronSledConfig`, and we do have at least
one of the legacy disks/datasets/zone ledgers, but not all three. This
should never happen (RSS provides all three), and if we're in that state
we don't know how to proceed.
2. During conversion of the three legacy configs into
`OmicronSledConfig`, we write the ledgered `OmicronSledConfig`, but
reading it back fails, and then we also fail to remove it. This would
leave us in a state where we have a bad conversion on disk, and again we
have no way to proceed.

I don't think it's really possible for either of those to happen in
practice, but wanted to call them out anyway.