[dns-server] Answers for the rack's zones should be authoritative
This makes most answers authoritative and changes tests to match. This
won't change rack-internal DNS behavior; we don't rely on `aa` in
answers for anything internal. External DNS servers will also now set
the `aa` bit for queries for silo domains, though, so our DNS answers
will look a bit more normal.

We're authoritative for `NXDomain` answers to names under the zones a
server is authoritative for that are not present, though we're not (yet)
sending an SOA record to indicate how long a client can cache the
negative reuslt.

We are *not* authoritative for `ServFail` answers even if they are for
names we're authoritative for. The `aa` bit here has no RFC-defined
meaning.