[external API] Add created time to webhook secret views (#8100)
When a webhook secret is created, it is assigned a UUID. In the API,
secrets are always referenced by UUIDs, and do not have names. This is
fine for programmatic access to secrets, such as when a receiver
implementation adds or rotates its secrets automatically. However, it's
less useful for a human user creating or deleting secrets in the UI, as
they would have to keep the various UUIDs straight, which can be
challenging. See discussion in #8083.

While we _could_ add names to the secret resource so that human users
can identify them more easily, this is not ideal either. Secrets, unlike
other resouces, are basically fungible: the only thing a user would
really ever do with them is rotate them, and they have no real semantic
meaning beyond being "the old secret" or "the new secret". Therefore,
giving them names basically just forces the user to come up with a
unique string like "secret1" and "secret2" and so on, which is a bit
unfortunate.

Instead, we should display the creation timestamps of secrets in the UI.
This way, users need not come up with unique names for every secret, but
they can distinguish them beyond their UUID, and the timestamp
communicates the one thing about webhook secrets that the user will
actually care about: is it "the old one" or "the new one"?

This branch updates the external API view for webhook secrets to add a
`time_created` field. The view type itself is renamed from
`WebhookSecretId` to just `WebhookSecret`, since it now contains more
than an ID.

Fixes #8083