[omdb] TUI for inspecting support bundles (#8034)
Adds the `omdb nexus sb inspect` command, which includes a dashboard
that can be used inspect support bundles.

- `dev-tools/omdb/src/bin/omdb/nexus.rs` includes the integration of
this command. Support bundles may be accessed by: UUID, path, or by
"most recently created".
- `dev-tools/support-bundle-reader-lib` exposes a dashboard for actually
running a TUI. This TUI acts on a dynamic `SupportBundleAccessor` trait,
which can implement a file-opening and file-reader interface, regardless
of where the support bundle is stored.
- `dev-tools/support-bundle-reader-lib/src/dashboard.rs` actually
implements the ratatui interface, which allows users to inspect parts of
the support bundle, and optionally dump the contents of the bundles to
stdout.


![image](https://github.com/user-attachments/assets/9b6e7df6-3123-497c-bc81-0e33b747bf5f)