[inventory] record caboose SIGN value (#8021)
Cabooses for RoT archives and RoT bootloader archives (but not SP
archives) contain a SIGN key that's a hash of the public key that the
RoT needs its software signed by. Software that's updating the RoT or
RoT bootloader needs this information to choose the appropriate Hubris
archive.

This commit collects the value during inventory, so Reconfigurator will
have it available.

Closes: https://github.com/oxidecomputer/omicron/issues/7914